Upload API & API structure refactor
